+++ src/spdr_posix.c
@@ -4,4 +4,4 @@
 #include <sys/types.h>
 #include <unistd.h>
 
-uint32_t uu_spdr_get_pid() { return getpid(); }
+uint32_t uu_spdr_get_pid() { return (uint32_t)getpid(); }
